pyproject.toml
[tool.poetry]
name = "parser-mts"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"
packages = [{include = "parser_mts"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
selenium = "^4.12.0"
openpyxl = "^3.1.2"
pandas = "^2.1.0"
django = "^4.2.5"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
aiohttp = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
